after adding images of workers
